requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.0
Django==3.2.9
django-cors-headers==3.10.0
django-debug-toolbar==3.2.2
django-environ==0.8.1
django-filter==21.1
django-redis==5.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
idna==3.3
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
pypinyin==0.44.0
pytz==2021.3
redis==3.5.3
requests==2.26.0
six==1.15.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.7
zhconv==1.4.3
